requirements.txt
pandas>=1.5.0
openpyxl>=3.0.0

# Build tools (cho tạo .exe)
pyinstaller>=5.0.0
pillow>=9.0.0  # Cho tạo icon
requests>=2.28.0
beautifulsoup4>=4.11.0
html5lib>=1.1
lxml>=4.9.0
chardet>=5.0.0
